Change behavior Excel file is copy when export is done

+++ src/windows/main_window.py
@@ -661,9 +661,6 @@ class MainWindow(updates.UpdateWatcher, QMainWindow):
                 _("Save Project..."),
                 recommended_path,
                 _("OpenShot Project (*.osp)"))[0]
-            
-            # Copy Excel
-            self.copyExcelFileToProject(file_path)
 
         if file_path:
             s.setDefaultPath(s.actionType.SAVE, file_path)
@@ -757,15 +754,16 @@ class MainWindow(updates.UpdateWatcher, QMainWindow):
             if ".osp" not in file_path:
                 file_path = "%s.osp" % file_path
 
-            # Copy Excel
-            self.copyExcelFileToProject(file_path)
-
             # Save new project
             self.save_project(file_path)
 
     def copyExcelFileToProject(self, pathProjectOpenshot) :
-        pathExcel = os.path.join(info.PATH, "SCE", "Export", "ExportSCE.xlsx")
+        nameExcel = "ExportSCE.xlsx"
+        pathExcel = os.path.join(info.PATH, "SCE", "Export", nameExcel)
         pathProject = os.path.dirname(pathProjectOpenshot)
+        if os.path.isfile(os.path.join(pathProject, nameExcel)) :
+            log.info(f"The Excel exists in the path : {pathProject}")
+            return
         log.info(f"Excel file will be copy from : {pathExcel} to {pathProject}")
         shutil.copy(pathExcel, pathProject)
 
@@ -900,11 +898,11 @@ class MainWindow(updates.UpdateWatcher, QMainWindow):
             file_path = get_app().project.current_filepath
             if not file_path:
                 return 
-        file_path = os.path.join(get_assets_path(file_path), "ExportSCE.csv")
+        filePathCSV = os.path.join(get_assets_path(file_path), "ExportSCE.csv")
         log.info(f"Export for SCE to {file_path}")
 
         try :
-            with open(file_path, 'w', newline="", encoding="latin-1")as file:
+            with open(filePathCSV, 'w', newline="", encoding="latin-1")as file:
                 csvWriter = csv.writer(file)
                 csvWriter.writerow(header)
                 csvWriter.writerows(data)
@@ -912,6 +910,10 @@ class MainWindow(updates.UpdateWatcher, QMainWindow):
         except :
             log.error("Not able to export Tags in the CSV file")
             QMessageBox.critical(self, "Error export SCE", "The export did work propely")
+            return
+        
+         # Copy Excel
+        self.copyExcelFileToProject(file_path)
